sync: testing

+++ include/storage.h
@@ -218,7 +218,7 @@ private:
 
 class QueueStore;
 struct __attribute__((__packed__)) MessagePageIndex {
-    uint64_t volatile page_idx;
+    uint32_t volatile page_idx;
     uint32_t prev_total_msg_size;
     uint16_t msg_size = 0;
     uint8_t volatile file_idx;
